User/password changes
Changed up the SQL. I think the professor explained not to have username as the primary but I may be wrong. He doesn't make his username as a primary key at least.. I also added sUsername and sPassword to the SalespersonHandler.

I'm not sure if we need to have the variables in the bo java classes, such as name, phone, email be written as sName, sPhone, sEmail for future files. I still get fuzzy on those details.